chore: run tests with jest
