+++ package.json
@@ -30,16 +30,19 @@
   "devDependencies": {
     "@rollup/plugin-typescript": "^8.2.0",
     "@types/express": "^4.17.1",
+    "@types/jest": "26.0.23",
     "@types/node": "15.12.1",
     "cross-env": "7.0.3",
     "express": "^4.17.1",
     "express-graphql": "^0.9.0",
     "graphql": "^15.1.0",
+    "jest": "27.0.4",
     "prettier": "2.3.1",
     "rollup": "^2.39.0",
+    "ts-jest": "27.0.2",
     "ts-node": "^10.0.0",
     "tslib": "^2.1.0",
-    "typescript": "^4.3.2"
+    "typescript": "4.3.2"
   },
   "peerDependencies": {
     "graphql": "^15.1.0"
@@ -48,6 +51,6 @@
     "build:cjs": "cross-env MODE=cjs rollup -c",
     "build:esm": "rollup -c",
     "build": "yarn build:cjs && yarn build:esm",
-    "test": "ts-node test/simple.ts"
+    "test": "jest"
   }
 }
