--- test/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../tsconfig.cjs.json",
-  "compilerOptions": {
-    "module": "commonjs"
-  }
-}
